update readme.md & docs

+++ driver/software/RF212b/phy/at86rf212b/src/phy_tx.c
@@ -243,8 +243,8 @@ void send_frame(PHY_CSMAMode_t csmaMode, bool txRetries)
     TRX_SLP_TR_HIGH();
 	trx_delay_micros(1);
     TRX_SLP_TR_LOW();
-	uint8_t context = 0;
-    trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, &context, 54000, SYS_TIME_SINGLE);
+	uint8_t context = 0U;
+    trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, (uintptr_t)&context, 54000, SYS_TIME_SINGLE);
     if(trxEIC_waitTimer == SYS_TIME_HANDLE_INVALID)
     {
             return;
@@ -293,7 +293,7 @@ void handle_tx_end_irq(bool underrun_occured)
 				TRX_SLP_TR_HIGH();
 				trx_delay_micros(1);
 				TRX_SLP_TR_LOW();
-				uint8_t context = 0;
+				uint8_t context = 0U;
     			trxEIC_waitTimer = SYS_TIME_CallbackRegisterUS(&trxEIC_waitTimerCb, (uintptr_t)&context, 54000, SYS_TIME_SINGLE);
     			if(trxEIC_waitTimer == SYS_TIME_HANDLE_INVALID)
     			{
